Move InviteSender to `:library:matrixui` module.
