+++ libraries/matrixui/src/main/res/values/localazy.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+  <string name="screen_invites_invited_you">"%1$s (%2$s) invited you"</string>
+</resources>
